Mark primitive display data tests RunnableOnService
