+++ sdks/java/core/src/test/java/org/apache/beam/sdk/transforms/display/DisplayDataEvaluator.java
@@ -41,10 +41,11 @@ public class DisplayDataEvaluator {
   private final PipelineOptions options;
 
   /**
-   * Create a new {@link DisplayDataEvaluator} using {@link TestPipeline#testingPipelineOptions()}.
+   * Create a new {@link DisplayDataEvaluator} using options returned from
+   * {@link #getDefaultOptions()}.
    */
   public static DisplayDataEvaluator create() {
-    return create(TestPipeline.testingPipelineOptions());
+    return create(getDefaultOptions());
   }
 
   /**
@@ -52,6 +53,13 @@ public class DisplayDataEvaluator {
    */
   public static DisplayDataEvaluator create(PipelineOptions pipelineOptions) {
     return new DisplayDataEvaluator(pipelineOptions);
+  }
+
+  /**
+   * The default {@link PipelineOptions} which will be used by {@link #create()}.
+   */
+  public static PipelineOptions getDefaultOptions() {
+    return TestPipeline.testingPipelineOptions();
   }
 
   private DisplayDataEvaluator(PipelineOptions options) {
